pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "recon-progressive"
version = "0.2.0"
description = "Modular reconnaissance tool with passive and active modules"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["recon", "security", "nmap", "ct", "dns", "whois"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyyaml>=5.4",
    "rich>=12.0.0"
]

[project.scripts]
recon-progressive = "recon_progressive.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/recon-progressive"
Repository = "https://github.com/yourusername/recon-progressive.git"
BugTracker = "https://github.com/yourusername/recon-progressive/issues"

[tool.setuptools.packages.find]
where = ["src"]